3-alloc_grid.c

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -8,21 +8,22 @@
 int **alloc_grid(int width, int height)
 {
 	int **arr;
-	size_t i, j;
+	int i, j;
+
 	if (width <= 0 || height <= 0)
 		return (NULL);
 	arr = (int **)malloc(sizeof(int) * (width));
 	if (arr == NULL)
 		return (NULL);
-	for (i = 0; i < (size_t)width; i++)
+	for (i = 0; i < width; i++)
 	{
 		arr[i] = (int *)malloc(height * sizeof(int));
 		if (arr[i] == NULL)
 			return (NULL);
 	}
-	for (i = 0; i < (size_t)height; i++)
+	for (i = 0; i < height; i++)
 	{
-		for (j = 0; j < (size_t)width; j++)
+		for (j = 0; j < width; j++)
 		{
 			arr[i][j] = 0;
 		}
